Refactored task service. Still need to remove semver
