+++ src/vs/workbench/contrib/tasks/browser/runAutomaticTasks.ts
