+++ src/vs/workbench/workbench.main.ts
@@ -268,7 +268,10 @@ import 'vs/workbench/contrib/terminal/browser/terminalPanel';
 import 'vs/workbench/contrib/relauncher/electron-browser/relauncher.contribution';
 
 // Tasks
-import 'vs/workbench/contrib/tasks/electron-browser/task.contribution';
+import 'vs/workbench/contrib/tasks/browser/task.contribution';
+import { TaskService } from 'vs/workbench/contrib/tasks/electron-browser/taskService';
+import { ITaskService } from 'vs/workbench/contrib/tasks/common/taskService';
+registerSingleton(ITaskService, TaskService, true);
 
 // Remote
 import 'vs/workbench/contrib/remote/electron-browser/remote.contribution';
